Changed Jira to be a plugin and moved appropriate Jira specific js to it. Built some plugin interfaces; i will change and built more interfaces that the plugins need to extend from so that we can have consistent metrics with any system that is plugged in
