+++ plugins/Jira/services/config.js
