+++ tests/unit/services/JiraPlugin/SprintCommitment.test.js
@@ -1,7 +1,7 @@
 import test from 'ava';
-import { JiraRest } from '../../../../services/adapters/jiraRest.js';
+import { JiraRest } from '../../../../plugins/Jira/services/jiraRest.js';
 import { ActoValidator } from '../../../../services/validators/ActoValidator.js'
-import { SprintCommitment } from '../../../../services/jira/sprintCommitment.js';
+import { SprintCommitment } from '../../../../plugins/Jira/services/SprintCommitment.js';
 
 test('SprintCommitment should set the jr property if an instance of JiraRest is passed', t => {
     const jr = new JiraRest(new ActoValidator());
@@ -24,6 +24,12 @@ test('SprintCommitment should fetch an active sprint on board 167', async t => {
     // Only pass the boardId, the rest is handled by the getSprints method
     const res = await SP.getSprints(167);
     t.true(res.values.length > 0);
+
+    const activeSprint = res.values[0];
+    t.is(activeSprint.state, 'active')
+    t.is(activeSprint.id, 982)
+    t.is(activeSprint.self, "https://actocloud.atlassian.net/rest/agile/1.0/sprint/982")
+    t.is(activeSprint.originBoardId, 167)
 });
 
 test('getIssuesInSprint(sprintId) should return the issues in the active sprint for board 167', async t => {
@@ -35,4 +41,11 @@ test('getIssuesInSprint(sprintId) should return the issues in the active sprint 
     const sprintId = res.values[0].id;
     const issues = await SP.getIssuesInSprint(sprintId, {})
     t.true(issues.length > 0);
+
+    // Validate the structure of the first issue
+    const firstIssue = issues[0];
+    t.true(typeof firstIssue.id === 'string');
+    t.true(typeof firstIssue.key === 'string');
+    t.true(typeof firstIssue.fields === 'object');
+    t.truthy(firstIssue.fields.summary);
 })
